+++ src/features/profile/pages/ChangeEmail/ChangeEmail.tsx
@@ -1,23 +1,15 @@
-import { useNavigation } from '@react-navigation/native'
 import React, { useEffect, useState, useRef } from 'react'
 import { Platform, ScrollView, StyleProp, ViewStyle } from 'react-native'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
-import { useMutation } from 'react-query'
 import styled, { useTheme } from 'styled-components/native'
 import { v4 as uuidv4 } from 'uuid'
 
-import { api } from 'api/api'
-import { ApiError } from 'api/apiHelpers'
 import { isLongEnough } from 'features/auth/components/PasswordSecurityRules'
-import { UseNavigationType } from 'features/navigation/RootNavigator/types'
-import { getTabNavConfig } from 'features/navigation/TabBar/helpers'
 import { AlreadyChangedEmailDisclaimer } from 'features/profile/components/Disclaimers/AlreadyChangedEmailDisclaimer'
 import { ChangeEmailDisclaimer } from 'features/profile/components/Disclaimers/ChangeEmailDisclaimer'
-import { CHANGE_EMAIL_ERROR_CODE } from 'features/profile/enums'
+import { useChangeEmailMutation } from 'features/profile/helpers/useChangeEmailMutation'
 import { useCheckHasCurrentEmailChange } from 'features/profile/helpers/useCheckHasCurrentEmailChange'
 import { useValidateEmail } from 'features/profile/helpers/useValidateEmail'
-import { ChangeEmailRequest } from 'features/profile/types'
-import { analytics } from 'libs/firebase/analytics'
 import { useSafeState } from 'libs/hooks'
 import { theme } from 'theme'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
@@ -27,7 +19,6 @@ import { EmailInput } from 'ui/components/inputs/EmailInput/EmailInput'
 import { InputError } from 'ui/components/inputs/InputError'
 import { PasswordInput } from 'ui/components/inputs/PasswordInput'
 import { useForHeightKeyboardEvents } from 'ui/components/keyboard/useKeyboardEvents'
-import { SNACK_BAR_TIME_OUT, useSnackBarContext } from 'ui/components/snackBar/SnackBarContext'
 import { useEnterKeyAction } from 'ui/hooks/useEnterKeyAction'
 import { getSpacing, Spacer } from 'ui/theme'
 
@@ -37,30 +28,12 @@ export function ChangeEmail() {
   const [password, setPassword] = useSafeState('')
   const { emailErrorMessage, isEmailValid } = useValidateEmail(email)
   const [passwordErrorMessage, setPasswordErrorMessage] = useSafeState<string | null>(null)
-  const { navigate } = useNavigation<UseNavigationType>()
-  const { showSuccessSnackBar, showErrorSnackBar } = useSnackBarContext()
   const { hasCurrentEmailChange } = useCheckHasCurrentEmailChange()
 
   const passwordInputErrorId = uuidv4()
   const emailInputErrorId = uuidv4()
 
-  const { mutate: changeEmail, isLoading } = useMutation(
-    (body: ChangeEmailRequest) => api.postnativev1profileupdateEmail(body),
-    {
-      onSuccess: () => {
-        showSuccessSnackBar({
-          message:
-            'E-mail envoyé\u00a0! Tu as 24h pour activer ta nouvelle adresse. Si tu ne le trouves pas, pense à vérifier tes spams.',
-          timeout: SNACK_BAR_TIME_OUT,
-        })
-        navigateToProfile()
-        analytics.logSaveNewMail()
-      },
-      onError: (error: ApiError | unknown) => {
-        onEmailChangeError((error as ApiError)?.content?.code)
-      },
-    }
-  )
+  const { changeEmail, isLoading } = useChangeEmailMutation({ setPasswordErrorMessage })
 
   useEffect(() => {
     removePasswordError()
@@ -72,24 +45,9 @@ export function ChangeEmail() {
   const { bottom } = useSafeAreaInsets()
   useForHeightKeyboardEvents(setKeyboardHeight)
 
-  const onEmailChangeError = (errorCode?: string) => {
-    errorCode && analytics.logErrorSavingNewEmail(errorCode)
-    if (errorCode === CHANGE_EMAIL_ERROR_CODE.INVALID_PASSWORD) {
-      setPasswordErrorMessage('Mot de passe incorrect')
-    } else {
-      showErrorSnackBar({
-        message:
-          'Une erreur s’est produite pendant la modification de ton e-mail. Réessaie plus tard.',
-        timeout: SNACK_BAR_TIME_OUT,
-      })
-    }
-  }
-
   const removePasswordError = () => {
     setPasswordErrorMessage(null)
   }
-
-  const navigateToProfile = () => navigate(...getTabNavConfig('Profile'))
 
   const submitEmailChange = () => {
     changeEmail({ email, password })
